fix: renamed imports from catalog core submodule
